requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
docker==6.0.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
multidict==6.0.4
openai==0.27.4
packaging==23.1
pycodestyle==2.10.0
python-git==2018.2.1
regex==2023.3.23
requests==2.28.2
Send2Trash==1.8.0
smmap==5.0.0
tiktoken==0.3.3
tqdm==4.65.0
urllib3==1.26.15
websocket-client==1.5.1
yarl==1.8.2
